requirements.txt
streamlit
google-api-python-client
google-generativeai
pytrends
pandas